Update Clip.cs
Duration can't be an integer as decimal values are a possible response and will throw a JsonReaderException

+++ TwitchLib.Api.Helix.Models/Clips/GetClips/Clip.cs
@@ -29,6 +29,6 @@ namespace TwitchLib.Api.Helix.Models.Clips.GetClips
         [JsonProperty(PropertyName = "thumbnail_url")]
         public string ThumbnailUrl { get; protected set; }
         [JsonProperty(PropertyName = "duration")]
-        public int Duration { get; protected set; }
+        public double Duration { get; protected set; }
     }
 }
